+++ source/rooms/artifactguardian-royalslime.js
@@ -3,6 +3,7 @@ const { generateCombatRoomBuilder } = require("../util/messageComponentUtil");
 
 module.exports = new RoomTemplate("A Slimy Throneroom",
 	"@{adventure}",
+	"Artifact Guardian",
 	"A long hall of wavy mirrors sits silently between the party and the door... until a bunch of shadows step out of the mirror and attack the party!",
 	[
 		new ResourceTemplate("3", "internal", "levelsGained"),