+++ source/rooms/event-freegoldonfire.js
@@ -5,6 +5,7 @@ const { SAFE_DELIMITER } = require("../constants");
 
 module.exports = new RoomTemplate("Free Gold?",
 	"Fire",
+	"Event",
 	"A large pile of gold sits quietly in the middle of the room, seemingly alone.",
 	[
 		new ResourceTemplate("300", "internal", "gold")