+++ source/rooms/restsite-trainingdummy.js
@@ -5,6 +5,7 @@ const { generateRoutingRow } = require("../util/messageComponentUtil");
 
 module.exports = new RoomTemplate("Rest Site: Training Dummy",
 	"@{adventure}",
+	"Rest Site",
 	"The room contains a campfire and a training dummy.",
 	[
 		new ResourceTemplate("n", "internal", "roomAction")